+++ src/core/CMakeLists.txt
@@ -46,6 +46,7 @@ set(SRCS
             hle/kernel/client_session.cpp
             hle/kernel/event.cpp
             hle/kernel/handle_table.cpp
+            hle/kernel/hle_ipc.cpp
             hle/kernel/kernel.cpp
             hle/kernel/memory.cpp
             hle/kernel/mutex.cpp
@@ -239,6 +240,7 @@ set(HEADERS
             hle/kernel/errors.h
             hle/kernel/event.h
             hle/kernel/handle_table.h
+            hle/kernel/hle_ipc.h
             hle/kernel/kernel.h
             hle/kernel/memory.h
             hle/kernel/mutex.h
