+++ src/core/hle/kernel/client_session.cpp
@@ -5,6 +5,8 @@
 #include "common/assert.h"
 
 #include "core/hle/kernel/client_session.h"
+#include "core/hle/kernel/errors.h"
+#include "core/hle/kernel/hle_ipc.h"
 #include "core/hle/kernel/server_session.h"
 
 namespace Kernel {
