+++ src/core/hle/kernel/hle_ipc.cpp
@@ -1,0 +1,22 @@
+// Copyright 2017 Citra Emulator Project
+// Licensed under GPLv2 or any later version
+// Refer to the license.txt file included.
+
+#include <boost/range/algorithm_ext/erase.hpp>
+#include "common/assert.h"
+#include "common/common_types.h"
+#include "core/hle/kernel/hle_ipc.h"
+#include "core/hle/kernel/kernel.h"
+#include "core/hle/kernel/server_session.h"
+
+namespace Kernel {
+
+void SessionRequestHandler::ClientConnected(SharedPtr<ServerSession> server_session) {
+    connected_sessions.push_back(server_session);
+}
+
+void SessionRequestHandler::ClientDisconnected(SharedPtr<ServerSession> server_session) {
+    boost::range::remove_erase(connected_sessions, server_session);
+}
+
+} // namespace Kernel
