+++ src/core/hle/kernel/server_session.h
@@ -12,7 +12,6 @@
 #include "core/hle/kernel/session.h"
 #include "core/hle/kernel/wait_object.h"
 #include "core/hle/result.h"
-#include "core/hle/service/service.h"
 #include "core/memory.h"
 
 namespace Kernel {
@@ -20,6 +19,7 @@ namespace Kernel {
 class ClientSession;
 class ClientPort;
 class ServerSession;
+class SessionRequestHandler;
 class Thread;
 
 /**
@@ -56,7 +56,7 @@ public:
      */
     static SessionPair CreateSessionPair(
         const std::string& name = "Unknown",
-        std::shared_ptr<Service::SessionRequestHandler> hle_handler = nullptr,
+        std::shared_ptr<SessionRequestHandler> hle_handler = nullptr,
         SharedPtr<ClientPort> client_port = nullptr);
 
     /**
@@ -72,7 +72,7 @@ public:
     std::string name;                ///< The name of this session (optional)
     bool signaled;                   ///< Whether there's new data available to this ServerSession
     std::shared_ptr<Session> parent; ///< The parent session, which links to the client endpoint.
-    std::shared_ptr<Service::SessionRequestHandler>
+    std::shared_ptr<SessionRequestHandler>
         hle_handler; ///< This session's HLE request handler (optional)
 
 private:
@@ -87,8 +87,7 @@ private:
      * @return The created server session
      */
     static ResultVal<SharedPtr<ServerSession>> Create(
-        std::string name = "Unknown",
-        std::shared_ptr<Service::SessionRequestHandler> hle_handler = nullptr);
+        std::string name = "Unknown", std::shared_ptr<SessionRequestHandler> hle_handler = nullptr);
 };
 
 /**
